+++ resources/android/src/com/victorycodedev/plugins/toastkit/model/ToastKitModels.kt
@@ -34,6 +34,7 @@ internal data class ToastKitConfiguration(
     val uniqueKey: String?,
     val nativeIos: Boolean,
     val nativeAndroid: Boolean,
+    val opacity: Float?,
     val message: String,
     val title: String?,
     val variant: String,